Start work on some basic built-in sampling policies
