+++ src/SerilogTracing/Samplers/AlwaysRecordedSampler.cs
@@ -1,0 +1,31 @@
+// Copyright © SerilogTracing Contributors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System.Diagnostics;
+
+namespace SerilogTracing.Samplers;
+
+/// <summary>
+/// Records all traces.
+/// </summary>
+static class AlwaysRecordedSampler
+{
+    /// <summary>
+    /// Create a sampling delegate that records all traces.
+    /// </summary>
+    public static SampleActivity<ActivityContext> Create()
+    {
+        return (ref ActivityCreationOptions<ActivityContext> _) => ActivitySamplingResult.AllDataAndRecorded;
+    }
+}
